@@ -4,12 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Beey.Api.DTO
+namespace Beey.Api.DTO;
+
+public class MonthlyTranscriptionLogItem
 {
-    public class MonthlyTranscriptionLogItem
-    {
-        public int Year { get; set; }
-        public int Month { get; set; }
-        public decimal Minutes { get; set; }
-    }
+    public int Year { get; set; }
+    public int Month { get; set; }
+    public decimal Minutes { get; set; }
 }
