@@ -3,21 +3,20 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Beey.Api.Rest
+namespace Beey.Api.Rest;
+
+public abstract class BaseAuthApi<TApi> : BaseApi<TApi> where TApi : BaseAuthApi<TApi>
 {
-    public abstract class BaseAuthApi<TApi> : BaseApi<TApi> where TApi : BaseAuthApi<TApi>
+    public LoginToken? Token { get; set; }
+
+    public BaseAuthApi(string url) : base(url)
     {
-        public LoginToken? Token { get; set; }
+    }
 
-        public BaseAuthApi(string url) : base(url)
-        {
-        }
+    internal override RestRequestBuilder CreateBuilder()
+    {
+        if (Token == null) { throw new UnauthorizedAccessException(); }
 
-        internal override RestRequestBuilder CreateBuilder()
-        {
-            if (Token == null) { throw new UnauthorizedAccessException(); }
-
-            return base.CreateBuilder().AddHeader("Authorization", Token.Token);
-        }
+        return base.CreateBuilder().AddHeader("Authorization", Token.Token);
     }
 }
