@@ -6,20 +6,19 @@ using System.Net;
 using System.Text;
 
 
-namespace Beey.Api
+namespace Beey.Api;
+
+internal class Utility
 {
-    internal class Utility
+    internal static void LogApiException(Exception ex, Logging.ILog logger)
     {
-        internal static void LogApiException(Exception ex, Logging.ILog logger)
+        if (ex is WebSocketClosedException wsEx && wsEx.CloseStatus.HasValue)
         {
-            if (ex is WebSocketClosedException wsEx && wsEx.CloseStatus.HasValue)
-            {
-                logger.Log(Logging.LogLevel.Error, () => $"Error in Beey API, WebSocket closed ({wsEx.CloseStatus?.ToString()}) with message '{wsEx.Message}'.", wsEx);
-            }
-            else
-            {
-                logger.Log(Logging.LogLevel.Error, () => $"Error in Beey API: '{ex.Message}'", ex);
-            }           
+            logger.Log(Logging.LogLevel.Error, () => $"Error in Beey API, WebSocket closed ({wsEx.CloseStatus?.ToString()}) with message '{wsEx.Message}'.", wsEx);
         }
+        else
+        {
+            logger.Log(Logging.LogLevel.Error, () => $"Error in Beey API: '{ex.Message}'", ex);
+        }           
     }
 }
