@@ -10,80 +10,79 @@ using Beey.Api.Rest;
 using System.Net;
 using Newtonsoft.Json.Linq;
 
-namespace Beey.Client
+namespace Beey.Client;
+
+public partial class BeeyClient
 {
-    public partial class BeeyClient
+    public async Task<IEnumerable<string>> GetTagsAsync(int id,
+        CancellationToken cancellationToken = default)
     {
-        public async Task<IEnumerable<string>> GetTagsAsync(int id,
-            CancellationToken cancellationToken = default)
+        this.RequireAuthorization();
+
+        var policy = CreateHttpAsyncUnauthorizedPolicy<string>();
+        var tags = (await policy.ExecuteAsync(async (ctx, c) =>
         {
-            this.RequireAuthorization();
+            return await ProjectApi.GetTagsAsync(id, c);
+        }, CreatePollyContext(cancellationToken), cancellationToken));
 
-            var policy = CreateHttpAsyncUnauthorizedPolicy<string>();
-            var tags = (await policy.ExecuteAsync(async (ctx, c) =>
-            {
-                return await ProjectApi.GetTagsAsync(id, c);
-            }, CreatePollyContext(cancellationToken), cancellationToken));
+        var jTags = JArray.Parse(tags);
+        return jTags.Select(t => t.Value<string>());
+    }
+    public async Task<Project> AddTagAsync(int id, long accessToken, string tag,
+        CancellationToken cancellationToken = default)
+    {
+        this.RequireAuthorization();
 
-            var jTags = JArray.Parse(tags);
-            return jTags.Select(t => t.Value<string>());
-        }
-        public async Task<Project> AddTagAsync(int id, long accessToken, string tag,
-            CancellationToken cancellationToken = default)
+        var policy = CreateHttpAsyncUnauthorizedPolicy<Project>();
+        return (await policy.ExecuteAsync(async (ctx, c) =>
         {
-            this.RequireAuthorization();
+            return await ProjectApi.AddTagAsync(id, accessToken, tag, c);
+        }, CreatePollyContext(cancellationToken), cancellationToken));
+    }
+    public async Task<Project> DeleteTagAsync(int id, long accessToken, string tag,
+        CancellationToken cancellationToken = default)
+    {
+        this.RequireAuthorization();
 
-            var policy = CreateHttpAsyncUnauthorizedPolicy<Project>();
-            return (await policy.ExecuteAsync(async (ctx, c) =>
-            {
-                return await ProjectApi.AddTagAsync(id, accessToken, tag, c);
-            }, CreatePollyContext(cancellationToken), cancellationToken));
-        }
-        public async Task<Project> DeleteTagAsync(int id, long accessToken, string tag,
-            CancellationToken cancellationToken = default)
+        var policy = CreateHttpAsyncUnauthorizedPolicy<Project>();
+        return (await policy.ExecuteAsync(async (ctx, c) =>
         {
-            this.RequireAuthorization();
+            return await ProjectApi.RemoveTagAsync(id, accessToken, tag, c);
+        }, CreatePollyContext(cancellationToken), cancellationToken));
+    }
 
-            var policy = CreateHttpAsyncUnauthorizedPolicy<Project>();
-            return (await policy.ExecuteAsync(async (ctx, c) =>
-            {
-                return await ProjectApi.RemoveTagAsync(id, accessToken, tag, c);
-            }, CreatePollyContext(cancellationToken), cancellationToken));
-        }
+    public async Task<ProjectMetadata> GetMetadataAsync(int id, string key,
+       CancellationToken cancellationToken)
+    {
+        this.RequireAuthorization();
 
-        public async Task<ProjectMetadata> GetMetadataAsync(int id, string key,
-           CancellationToken cancellationToken)
+        var policy = CreateHttpAsyncUnauthorizedPolicy<ProjectMetadata>();
+        return await policy.ExecuteAsync(async (ctx, c) =>
         {
-            this.RequireAuthorization();
+            return await ProjectApi.GetMetadataAsync(id, key, c);
+        }, CreatePollyContext(cancellationToken), cancellationToken);
+    }
+    public async Task<Project> AddMetadataAsync(int id, long accessToken,
+        string key, string value,
+       CancellationToken cancellationToken)
+    {
+        this.RequireAuthorization();
 
-            var policy = CreateHttpAsyncUnauthorizedPolicy<ProjectMetadata>();
-            return await policy.ExecuteAsync(async (ctx, c) =>
-            {
-                return await ProjectApi.GetMetadataAsync(id, key, c);
-            }, CreatePollyContext(cancellationToken), cancellationToken);
-        }
-        public async Task<Project> AddMetadataAsync(int id, long accessToken,
-            string key, string value,
-           CancellationToken cancellationToken)
+        var policy = CreateHttpAsyncUnauthorizedPolicy<Project>();
+        return (await policy.ExecuteAsync(async (ctx, c) =>
         {
-            this.RequireAuthorization();
+            return await ProjectApi.AddMetadataAsync(id, accessToken, key, value, c);
+        }, CreatePollyContext(cancellationToken), cancellationToken));
+    }
+    public async Task<Project> RemoveMetadataAsync(int id, long accessToken, string key,
+       CancellationToken cancellationToken)
+    {
+        this.RequireAuthorization();
 
-            var policy = CreateHttpAsyncUnauthorizedPolicy<Project>();
-            return (await policy.ExecuteAsync(async (ctx, c) =>
-            {
-                return await ProjectApi.AddMetadataAsync(id, accessToken, key, value, c);
-            }, CreatePollyContext(cancellationToken), cancellationToken));
-        }
-        public async Task<Project> RemoveMetadataAsync(int id, long accessToken, string key,
-           CancellationToken cancellationToken)
+        var policy = CreateHttpAsyncUnauthorizedPolicy<Project>();
+        return (await policy.ExecuteAsync(async (ctx, c) =>
         {
-            this.RequireAuthorization();
-
-            var policy = CreateHttpAsyncUnauthorizedPolicy<Project>();
-            return (await policy.ExecuteAsync(async (ctx, c) =>
-            {
-                return await ProjectApi.RemoveMetadataAsync(id, accessToken, key, c);
-            }, CreatePollyContext(cancellationToken), cancellationToken));
-        }
+            return await ProjectApi.RemoveMetadataAsync(id, accessToken, key, c);
+        }, CreatePollyContext(cancellationToken), cancellationToken));
     }
 }
