@@ -7,69 +7,68 @@ using System.Threading;
 using System.Threading.Tasks;
 using TranscriptionCore;
 
-namespace Beey.Client
+namespace Beey.Client;
+
+public partial class BeeyClient
 {
-    public partial class BeeyClient
+    public async Task<Listing<Speaker>> ListSpeakersAsync(int count, int skip = 0, string? search = null,
+        CancellationToken cancellationToken = default)
     {
-        public async Task<Listing<Speaker>> ListSpeakersAsync(int count, int skip = 0, string? search = null,
-            CancellationToken cancellationToken = default)
+        this.RequireAuthorization();
+
+        var policy = CreateHttpAsyncUnauthorizedPolicy<Listing<Speaker>>();
+        return (await policy.ExecuteAsync(async (ctx, c) =>
         {
-            this.RequireAuthorization();
+            return await SpeakerApi.ListAsync(count, skip, search, c);
+        }, CreatePollyContext(cancellationToken), cancellationToken));
+    }
 
-            var policy = CreateHttpAsyncUnauthorizedPolicy<Listing<Speaker>>();
-            return (await policy.ExecuteAsync(async (ctx, c) =>
-            {
-                return await SpeakerApi.ListAsync(count, skip, search, c);
-            }, CreatePollyContext(cancellationToken), cancellationToken));
-        }
+    public async Task<Speaker> GetSpeakerAsync(string dbId,
+        CancellationToken cancellationToken = default)
+    {
+        this.RequireAuthorization();
 
-        public async Task<Speaker> GetSpeakerAsync(string dbId,
-            CancellationToken cancellationToken = default)
+        var policy = CreateHttpAsyncUnauthorizedPolicy<Speaker>();
+        return (await policy.ExecuteAsync(async (ctx, c) =>
         {
-            this.RequireAuthorization();
+            return await SpeakerApi.GetAsync(dbId, c);
+        }, CreatePollyContext(cancellationToken), cancellationToken));
+    }
 
-            var policy = CreateHttpAsyncUnauthorizedPolicy<Speaker>();
-            return (await policy.ExecuteAsync(async (ctx, c) =>
-            {
-                return await SpeakerApi.GetAsync(dbId, c);
-            }, CreatePollyContext(cancellationToken), cancellationToken));
-        }
+    public async Task<Speaker> CreateSpeakerAsync(Speaker speaker,
+        CancellationToken cancellationToken = default)
+    {
+        this.RequireAuthorization();
 
-        public async Task<Speaker> CreateSpeakerAsync(Speaker speaker,
-            CancellationToken cancellationToken = default)
+        var policy = CreateHttpAsyncUnauthorizedPolicy<Speaker>();
+        return (await policy.ExecuteAsync(async (ctx, c) =>
         {
-            this.RequireAuthorization();
+            return await SpeakerApi.CreateAsync(speaker, c);
+        }, CreatePollyContext(cancellationToken), cancellationToken));
+    }
 
-            var policy = CreateHttpAsyncUnauthorizedPolicy<Speaker>();
-            return (await policy.ExecuteAsync(async (ctx, c) =>
-            {
-                return await SpeakerApi.CreateAsync(speaker, c);
-            }, CreatePollyContext(cancellationToken), cancellationToken));
-        }
+    public async Task UpdateSpeakerAsync(Speaker speaker,
+        CancellationToken cancellationToken = default)
+    {
+        this.RequireAuthorization();
 
-        public async Task UpdateSpeakerAsync(Speaker speaker,
-            CancellationToken cancellationToken = default)
+        var policy = CreateHttpAsyncUnauthorizedPolicy<bool>();
+        await policy.ExecuteAsync(async (ctx, c) =>
         {
-            this.RequireAuthorization();
+            await SpeakerApi.UpdateAsync(speaker, c);
+            return true;
+        }, CreatePollyContext(cancellationToken), cancellationToken);
+    }
 
-            var policy = CreateHttpAsyncUnauthorizedPolicy<bool>();
-            await policy.ExecuteAsync(async (ctx, c) =>
-            {
-                await SpeakerApi.UpdateAsync(speaker, c);
-                return true;
-            }, CreatePollyContext(cancellationToken), cancellationToken);
-        }
+    public async Task<bool> DeleteSpeakerAsync(string dbId,
+        CancellationToken cancellationToken = default)
+    {
+        this.RequireAuthorization();
 
-        public async Task<bool> DeleteSpeakerAsync(string dbId,
-            CancellationToken cancellationToken = default)
+        var policy = CreateHttpAsyncUnauthorizedPolicy<bool>();
+        return (await policy.ExecuteAsync(async (ctx, c) =>
         {
-            this.RequireAuthorization();
-
-            var policy = CreateHttpAsyncUnauthorizedPolicy<bool>();
-            return (await policy.ExecuteAsync(async (ctx, c) =>
-            {
-                return await SpeakerApi.DeleteAsync(dbId, c);
-            }, CreatePollyContext(cancellationToken), cancellationToken));
-        }
+            return await SpeakerApi.DeleteAsync(dbId, c);
+        }, CreatePollyContext(cancellationToken), cancellationToken));
     }
 }
