@@ -3,74 +3,73 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Beey.Client
+namespace Beey.Client;
+
+public class TryResult
 {
-    public class TryResult
+    public Exception? Exception { get; private set; }
+    public bool IsSuccess { get => Exception == null; }
+
+    public TryResult() { }
+    public TryResult(Exception ex)
     {
-        public Exception? Exception { get; private set; }
-        public bool IsSuccess { get => Exception == null; }
+        Exception = ex;
+    }
 
-        public TryResult() { }
-        public TryResult(Exception ex)
-        {
-            Exception = ex;
-        }
+    public static implicit operator bool(TryResult tryResult)
+    {
+        return tryResult.IsSuccess;
+    }
+}
 
-        public static implicit operator bool(TryResult tryResult)
+public class TryValueResult<T> : TryResult
+{
+    // hack to get rid of nullable reference warnings
+    private object? value;
+    public T Value
+    {
+        get
         {
-            return tryResult.IsSuccess;
+            return IsSuccess ? (T)value! : throw new ArgumentNullException();
         }
     }
 
-    public class TryValueResult<T> : TryResult
+    public TryValueResult(T result) : base()
     {
-        // hack to get rid of nullable reference warnings
-        private object? value;
-        public T Value
-        {
-            get
-            {
-                return IsSuccess ? (T)value! : throw new ArgumentNullException();
-            }
-        }
+        this.value = result;
+    }
 
-        public TryValueResult(T result) : base()
-        {
-            this.value = result;
-        }
+    public TryValueResult(Exception ex) : base(ex)
+    {
+        this.value = default;
+    }
+}
 
-        public TryValueResult(Exception ex) : base(ex)
+public static class TaskExtensions
+{
+    public static async Task<TryResult> TryAsync(this Task task)
+    {
+        try
         {
-            this.value = default;
+            await task;
+            return new TryResult();
+        }
+        catch (Exception ex)
+        {
+            return new TryResult(ex);
         }
     }
 
-    public static class TaskExtensions
+    public static async Task<TryValueResult<TResult>> TryAsync<TResult>(this Task<TResult> task)
     {
-        public static async Task<TryResult> TryAsync(this Task task)
+        try
         {
-            try
-            {
-                await task;
-                return new TryResult();
-            }
-            catch (Exception ex)
-            {
-                return new TryResult(ex);
-            }
+            var result = await task;
+            return new TryValueResult<TResult>(result);
         }
-
-        public static async Task<TryValueResult<TResult>> TryAsync<TResult>(this Task<TResult> task)
+        catch (Exception ex)
         {
-            try
-            {
-                var result = await task;
-                return new TryValueResult<TResult>(result);
-            }
-            catch (Exception ex)
-            {
-                return new TryValueResult<TResult>(ex);
-            }
+            return new TryValueResult<TResult>(ex);
         }
     }
 }
