@@ -7,24 +7,23 @@ using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace DemoBeeyApp.Controllers
+namespace DemoBeeyApp.Controllers;
+
+[Route("/")]
+public class HomeController : Controller
 {
-    [Route("/")]
-    public class HomeController : Controller
+    private readonly ILogger<HomeController> _logger;
+    private readonly BeeyProxy beey;
+
+    public HomeController(ILogger<HomeController> logger, BeeyProxy beey)
     {
-        private readonly ILogger<HomeController> _logger;
-        private readonly BeeyProxy beey;
+        _logger = logger;
+        this.beey = beey;
+    }
 
-        public HomeController(ILogger<HomeController> logger, BeeyProxy beey)
-        {
-            _logger = logger;
-            this.beey = beey;
-        }
-
-        [HttpGet("/")]
-        public IActionResult Index()
-        {
-            return Json(beey);
-        }
+    [HttpGet("/")]
+    public IActionResult Index()
+    {
+        return Json(beey);
     }
 }
