@@ -10,12 +10,11 @@ using Xunit;
 // Need to turn off test parallelization so we can validate the run order
 [assembly: CollectionBehavior(DisableTestParallelization = true)]
 
-namespace XUnitTests
+namespace XUnitTests;
+
+public static class Configuration
 {
-    public static class Configuration
-    {
-        public static string BeeyUrl { get => "http://localhost:61497"; } 
-        public static string Email { get => "milos.kudelka@newtontech.cz"; }
-        public static string Password { get => "OVPgod"; }
-    }
+    public static string BeeyUrl { get => "http://localhost:61497"; } 
+    public static string Email { get => "milos.kudelka@newtontech.cz"; }
+    public static string Password { get => "OVPgod"; }
 }
