@@ -5,13 +5,12 @@ using System.Text;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace XUnitTests
+namespace XUnitTests;
+
+class TestCollectionOrderer : ITestCollectionOrderer
 {
-    class TestCollectionOrderer : ITestCollectionOrderer
+    public IEnumerable<ITestCollection> OrderTestCollections(IEnumerable<ITestCollection> testCollections)
     {
-        public IEnumerable<ITestCollection> OrderTestCollections(IEnumerable<ITestCollection> testCollections)
-        {
-            return testCollections.OrderBy(collection => collection.DisplayName);
-        }
+        return testCollections.OrderBy(collection => collection.DisplayName);
     }
 }
