@@ -2,16 +2,15 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace XUnitTests
-{
-    [AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
-    class TestPriorityAttribute : Attribute
-    {
-        public double Priority { get; private set; }
+namespace XUnitTests;
 
-        public TestPriorityAttribute(double priority)
-        {
-            Priority = priority;
-        }
+[AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
+class TestPriorityAttribute : Attribute
+{
+    public double Priority { get; private set; }
+
+    public TestPriorityAttribute(double priority)
+    {
+        Priority = priority;
     }
 }
